# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Response payload for successful device addition. The websocket_url is used to establish a connection that:\n- Receives media/images from the platform bridge\n- Handles login synchronization events\n- Receives existing active sessions if available\n- Closes when receiving nil data (indicating end of session or error)",
        "title": "ShortMesh API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "post": {
                "description": "Registers a new user and returns an access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creates a new user on the Matrix server",
                "parameters": [
                    {
                        "description": "User Registration",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ClientJsonRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created user",
                        "schema": {
                            "$ref": "#/definitions/main.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "User creation failed",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Authenticates a user and returns an access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Logs a user into the Matrix server",
                "parameters": [
                    {
                        "description": "Login Credentials",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ClientJsonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully logged in",
                        "schema": {
                            "$ref": "#/definitions/main.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Login failed",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/{platform}/devices": {
            "post": {
                "description": "Registers a new device connection for the specified platform and establishes a websocket connection.\nThe websocket connection will:\n- Receive media/images from the platform bridge\n- Handle login synchronization events\n- Send existing active sessions if available\n- Close connection when receiving nil data (indicating end of session or error)\nHere are various platforms supported:\n'wa' (for WhatsApp)\n'signal' (for Signal)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Adds a device for a given platform",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform Name",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Device Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ClientBridgeJsonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully added device and established websocket connection",
                        "schema": {
                            "$ref": "#/definitions/main.DeviceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid access token",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/{platform}/message/{contact}": {
            "post": {
                "description": "Sends a message to a contact through the specified platform bridge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Sends a message to a specified room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform Name",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Contact ID (E.164 phone number without the plus)",
                        "name": "contact",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ClientMessageJsonRequeset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message sent successfully",
                        "schema": {
                            "$ref": "#/definitions/main.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to send message",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ClientBridgeJsonRequest": {
            "description": "Request payload to bind a platform bridge to a user",
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "syt_YWxwaGE..."
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        },
        "main.ClientJsonRequest": {
            "description": "Request payload for user login or registration",
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "securepassword123"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        },
        "main.ClientMessageJsonRequeset": {
            "description": "Request payload to send a message to a room",
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "syt_YWxwaGE..."
                },
                "message": {
                    "type": "string",
                    "example": "Hello, world!"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        },
        "main.DeviceResponse": {
            "description": "Response payload for successful device addition. The websocket_url is used to establish a connection that: - Receives media/images from the platform bridge - Handles login synchronization events - Receives existing active sessions if available - Closes when receiving nil data (indicating end of session or error)",
            "type": "object",
            "properties": {
                "websocket_url": {
                    "type": "string",
                    "example": "ws://localhost:8080/ws/telegram/john_doe"
                }
            }
        },
        "main.ErrorResponse": {
            "description": "Response payload for error cases",
            "type": "object",
            "properties": {
                "details": {
                    "type": "string",
                    "example": "Username must be 3-32 characters"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
        "main.LoginResponse": {
            "description": "Response payload for successful login",
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "syt_YWxwaGE..."
                },
                "status": {
                    "type": "string",
                    "example": "logged in"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        },
        "main.MessageResponse": {
            "description": "Response payload for successful message sending",
            "type": "object",
            "properties": {
                "contact": {
                    "type": "string",
                    "example": "+1234567890"
                },
                "event_id": {
                    "type": "string",
                    "example": "$1234567890abcdef"
                },
                "message": {
                    "type": "string",
                    "example": "Hello, world!"
                },
                "status": {
                    "type": "string",
                    "example": "sent"
                }
            }
        }
    }
}